# swagger.json
{
  "openapi": "3.0.0",
      "info": {
        "title": "Work Wise API Server",
        "description": "The Work Wise API description",
        "version": "1.0.0",
        "contact": {
            "name": "The Work Wise API description",
            "email": "workwise.org@gmail.com",
            "url": "https://work-wise-server-side.onrender.com"
        }
    },
    "servers": [
        {
            "url": "https://work-wise-server-side.onrender.com"
          },
          {
            "url": "http://localhost:3000"
          }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
              "type": "http",
              "scheme": "bearer"
            },
            "noauthAuth": {
              "type": "http",
              "scheme": "noauth"
            }
          }
    },
    "tags": [
      {
        "name": "Users"
      }
],
  "paths": {
    "/": {
      "get": {
        "tags": [ "Users" ],
        "operationId": "AppController_root",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },

    "/users": {
      "post": {
        "tags": [ "Users" ],
        "summary": "User Signup",
        "operationId": "AuthController_addUsers",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "fullName":"",
                  "email":"",
                  "password": ""
                }
              }
            }
          }
        },   
        "responses": {
          "201": {
            "description": ""
          }
        }
      },
      "patch": {
        "operationId": "AuthController_updateMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "get": {
        "operationId": "AuthController_getUsers",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": [ "Users" ],
        "summary": "User Login",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "email":"",
                  "password": ""
                }
              }
            }
          }
        },   
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/users/google": {
      "get": {
        "tags": [ "Users" ],
        "operationId": "AuthController_googleAuth",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/users/google/callback": {
      "get": {
        "tags": [ "Users" ],
        "summary": "Google Signup",
        "operationId": "AuthController_googleAuthRedirect",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/users/verify": {
      "post": {
        "tags": [ "Users" ],
        "summary": "User Otp Verification",
        "operationId": "AuthController_verifyUsers",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "otpCode":""
                }
              }
            }
          }
        }, 
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/users/forgot-password": {
      "post": {
        "tags": [ "Users" ],
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "email":""
                }
              }
            }
          }
        },   
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/users/reset-password": {
      "post": {
        "tags": [ "Users" ],
        "operationId": "AuthController_resetPassword",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/users/change-password": {
      "post": {
        "tags": [ "Users" ],
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  }
}